+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes of I18N Translator will be documented here.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## [0.3.0] - Unreleased
+## [0.3.0] - 2026-06-17
 ### Added
 - Glossary Sidebar
 - Glossary Actions
@@ -193,6 +193,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Initialized
 - Main Page with a sample test
 
+[0.3.0]: https://github.com/ArsenTech/i18n-translator/releases/tag/v0.3.0
 [0.2.7]: https://github.com/ArsenTech/i18n-translator/releases/tag/v0.2.7
 [0.2.6]: https://github.com/ArsenTech/i18n-translator/releases/tag/v0.2.6
 [0.2.5]: https://github.com/ArsenTech/i18n-translator/releases/tag/v0.2.5